editoast: mv `map::layers` `models::layers`
First step of exploding the `map` module to isolate what is related to
the database. All the query setupn is now in models.
