+++ editoast/src/client/infra_commands.rs
@@ -12,8 +12,8 @@ use schemas::infra::RailJson;
 
 use crate::infra_cache::InfraCache;
 use crate::map;
-use crate::map::MapLayers;
 use crate::models::Infra;
+use crate::models::layers::MapLayers;
 use cache;
 use editoast_models::prelude::*;
