+++ editoast/src/views/test_app.rs
@@ -38,8 +38,8 @@ use url::Url;
 use crate::AppState;
 use crate::generated_data::speed_limit_tags_config::SpeedLimitTagIds;
 use crate::infra_cache::InfraCache;
-use crate::map::MapLayers;
 use crate::models::PgAuthDriver;
+use crate::models::layers::MapLayers;
 use crate::views::service_router;
 use fga::client::DEFAULT_OPENFGA_MAX_CHECKS_PER_BATCH_CHECK;
 use fga::client::DEFAULT_OPENFGA_MAX_TUPLES_PER_WRITE;
